+++ app/Http/Controllers/admin/UserController.php
@@ -7,41 +7,58 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-    public function index(){
-        $users = User::all()->where('role_as','!=',0);
-        return view('admin.user.index',compact('users'));
-        return DataTables::of($users)->make(true);
+    public function index()
+    {
+        if (Gate::allows('user.view'))
+        {
+            $users = User::all()->where('role_as','!=',0);
+            return view('admin.user.index',compact('users'));
+            return DataTables::of($users)->make(true);
+        } else {
+            abort(403, "You don't have permission to access");
+        }
     }
 
-    public function show(){
-        // $users = User::all()->where('role_as','!=',0);
-        // return DataTables::of($users, compact('users'))->make(true);
-        // // ->addColumn('role', function(User $user){
-        // //     $user = $user->roles->name;
-        // //     return $user;
-        // // })
-        // ->editColumn('action', function($user){
-        //     $btn = '
-        //     <div class="d-grid gap-2 d-md-block"> 
-        //         <a href='.route('show-edit-user',$user->id).' class="btn btn-sm btn-outline-info"><i class="material-icons">edit</i>Edit
-        //         </a>
-        //         <a href='.route('delete-user',$user->id).' onclick="return confirm(\'Are you sure?\')" class="btn btn-sm btn-outline-danger"><i class="material-icons">delete</i>
-        //         Delete
-        //         </a>
-        //     </div>';
-        //     return $btn;
-        // })
-        // ->rawColumns(['role', 'action'])
-        // ->make(true);
+    // public function show()
+    // {
+    //     // $users = User::all()->where('role_as','!=',0);
+    //     // return DataTables::of($users, compact('users'))->make(true);
+    //     // // ->addColumn('role', function(User $user){
+    //     // //     $user = $user->roles->name;
+    //     // //     return $user;
+    //     // // })
+    //     // ->editColumn('action', function($user){
+    //     //     $btn = '
+    //     //     <div class="d-grid gap-2 d-md-block"> 
+    //     //         <a href='.route('show-edit-user',$user->id).' class="btn btn-sm btn-outline-info"><i class="material-icons">edit</i>Edit
+    //     //         </a>
+    //     //         <a href='.route('delete-user',$user->id).' onclick="return confirm(\'Are you sure?\')" class="btn btn-sm btn-outline-danger"><i class="material-icons">delete</i>
+    //     //         Delete
+    //     //         </a>
+    //     //     </div>';
+    //     //     return $btn;
+    //     // })
+    //     // ->rawColumns(['role', 'action'])
+    //     // ->make(true);
+    // }
+
+    public function add()
+    {
+        if (Gate::allows('user.add'))
+        {
+            return view('admin.user.add');
+        } else {
+            abort(403, "You don't have permission to access");
+        }
     }
-    public function add(){
-        return view('admin.user.add');
-    }
-    public function store(Request $request){
+
+    public function store(Request $request)
+    {
         $user = new User;
         $user->name = $request->input('name');
         $user->email = $request->input('email');
@@ -52,11 +69,20 @@ class UserController extends Controller
         $user->save();
         return redirect('admin/user')->with('status','User Added Successfully');
     }
-    public function showedit($id){
-        $user = User::find($id);
-        return view('admin.user.edit',compact('user'));
+
+    public function showedit($id)
+    {
+        if (Gate::allows('user.edit'))
+        {
+            $user = User::find($id);
+            return view('admin.user.edit',compact('user'));
+        } else {
+            abort(403, "You don't have permission to access");
+        }
     }
-    public function edit(request $request,$id){
+
+    public function edit(request $request,$id)
+    {
         $user = User::find($id);
         $user->name = $request->input('name');
         $user->email = $request->input('email');
@@ -67,9 +93,15 @@ class UserController extends Controller
         return redirect('admin/user')->with('status','User Updated Successfully');
     }
 
-    public function delete($id){
-        $user = User::find($id);
-        $user->delete();
-        return redirect('admin/user')->with('status','User Deleted Successfully');
+    public function delete($id)
+    {
+        if (Gate::allows('user.delete'))
+        {
+            $user = User::find($id);
+            $user->delete();
+            return redirect('admin/user')->with('status','User Deleted Successfully');
+        } else {
+            abort(403, "You don't have permission to access");
+        }
     }
 }
